fixup! use existing MockRatisServer
