+++ hadoop-hdds/server-scm/src/test/java/org/apache/hadoop/hdds/scm/ha/MockSCMHAManager.java
@@ -125,84 +125,4 @@ public final class MockSCMHAManager implements SCMHAManager {
         RaftPeerId.valueOf("peer"), RaftGroupId.randomId()),
         null, new ArrayList<>());
   }
-
-  private static class MockRatisServer implements SCMRatisServer {
-
-    private Map<RequestType, Object> handlers =
-        new EnumMap<>(RequestType.class);
-
-    @Override
-    public void start() {
-    }
-
-    @Override
-    public void registerStateMachineHandler(final RequestType handlerType,
-                                            final Object handler) {
-      handlers.put(handlerType, handler);
-    }
-
-    @Override
-    public SCMRatisResponse submitRequest(final SCMRatisRequest request)
-        throws IOException {
-      final RaftGroupMemberId raftId = RaftGroupMemberId.valueOf(
-          RaftPeerId.valueOf("peer"), RaftGroupId.randomId());
-      RaftClientReply reply;
-      try {
-        final Message result = process(request);
-        return SCMRatisResponse.decode(new RaftClientReply(ClientId.randomId(),
-            raftId, 1L, true, result, null, 1L, null));
-      } catch (Exception ex) {
-        return SCMRatisResponse.decode(new RaftClientReply(ClientId.randomId(),
-            raftId, 1L, false, null,
-            new StateMachineException(raftId, ex), 1L, null));
-      }
-    }
-
-    private Message process(final SCMRatisRequest request)
-        throws Exception {
-      try {
-        final Object handler = handlers.get(request.getType());
-
-        if (handler == null) {
-          throw new IOException("No handler found for request type " +
-              request.getType());
-        }
-
-        final List<Class<?>> argumentTypes = new ArrayList<>();
-        for(Object args : request.getArguments()) {
-          argumentTypes.add(args.getClass());
-        }
-        final Object result = handler.getClass().getMethod(
-            request.getOperation(), argumentTypes.toArray(new Class<?>[0]))
-            .invoke(handler, request.getArguments());
-
-        return SCMRatisResponse.encode(result);
-      } catch (NoSuchMethodException | SecurityException ex) {
-        throw new InvalidProtocolBufferException(ex.getMessage());
-      } catch (InvocationTargetException e) {
-        final Exception targetEx = (Exception) e.getTargetException();
-        throw targetEx != null ? targetEx : e;
-      }
-    }
-
-    @Override
-    public RaftServer getServer() {
-      return null;
-    }
-
-    @Override
-    public RaftGroupId getRaftGroupId() {
-      return null;
-    }
-
-    @Override
-    public List<RaftPeer> getRaftPeers() {
-      return new ArrayList<>();
-    }
-
-    @Override
-    public void stop() {
-    }
-  }
-
 }
